Fix a bug in search
A measure was not mappend-ed were needed.

This bug was undetected by the testsuite, as there are no tests for search!
Consequently, this patch adds a test case and a QuickCheck property for
search. These tests fail without the bugfix. The tests are based on the
function lookupByIndexFromEnd, which led me to discover this bug in the first
place.